@@ -21,13 +21,13 @@
     "rc-menu": "^9.11.1",
     "rc-slider": "^9.7.5",
     "re-resizable": "^6.9.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-collapsible": "^2.10.0",
     "react-color": "^2.19.3",
     "react-copy-to-clipboard": "^5.0.4",
     "react-datepicker": "^4.16.0",
     "react-device-detect": "^2.2.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-ga4": "^2.0.0",
     "react-highlight-words": "^0.20.0",
     "react-icons": "^4.8.0",
