@@ -1,9 +1,10 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "./App";
 
 it("renders without crashing", () => {
-	const div = document.createElement("div");
-	ReactDOM.render(<App />, div);
-	ReactDOM.unmountComponentAtNode(div);
+  const div = document.createElement("div");
+  const root = createRoot(div); // createRoot(container!) if you use TypeScript
+  root.render(<App />);
+  root.unmount();
 });
