@@ -1,6 +1,6 @@
 // REACT
-import React, { Component, Fragment } from "react";
-import ReactDOM from "react-dom";
+import React, { Component, Fragment, useEffect } from "react";
+import { createRoot } from "react-dom/client";
 import { CSSTransition, TransitionGroup } from "react-transition-group";
 import { SubMenu, Item as MenuItem } from "rc-menu";
 
@@ -811,8 +811,7 @@ class MyMaps extends Component {
         </FloatingMenu>
       </Portal>
     );
-
-    ReactDOM.render(menu, document.getElementById("portal-root"));
+    window.portalRoot.render(menu);
   };
 
   onMenuItemClick = (action, item) => {
@@ -963,7 +962,6 @@ class MyMaps extends Component {
     var evtClone = Object.assign({}, evt);
     const feature = helpers.getFeatureFromGeoJSON(item.featureGeoJSON);
     const showCoordinates = feature.get("is_open_data");
-
     const menu = (
       <Portal>
         <FloatingMenu
@@ -1042,8 +1040,7 @@ class MyMaps extends Component {
         </FloatingMenu>
       </Portal>
     );
-
-    ReactDOM.render(menu, document.getElementById("portal-root"));
+    window.portalRoot.render(menu);
   };
 
   deleteSelected = (selected) => {
